8213411: JDK-8209189 incorrect for Big Endian (JVM crashes)

+++ src/hotspot/share/code/codeCache.hpp
@@ -192,7 +192,7 @@ class CodeCache : AllStatic {
     }
   };
   static void do_unloading(BoolObjectClosure* is_alive, bool unloading_occurred);
-  static uint16_t unloading_cycle() { return _unloading_cycle; }
+  static uint8_t unloading_cycle() { return _unloading_cycle; }
   static void increment_unloading_cycle();
   static void asserted_non_scavengable_nmethods_do(CodeBlobClosure* f = NULL) PRODUCT_RETURN;
   static void release_exception_cache(ExceptionCache* entry);
